テクスチャにtypeを追加した
