+++ src/js/resource/texture/configs/battery-number.ts
@@ -7,4 +7,4 @@ export const BatteryNumberTextureConfigs: TextureConfig[] = [
     id: TEXTURE_IDS.BATTERY_NUMBER,
     path: "battery-number/battery-number.webp",
   },
-];
+].map((t) => ({ ...t, type: "Shared" }));
