+++ src/js/resource/texture/configs/gai.ts
@@ -7,4 +7,4 @@ export const GaiTextureConfigs: TextureConfig[] = [
     id: TEXTURE_IDS.GAI_CUTIN,
     path: "pilot/gai/skill-cutin.webp",
   },
-];
+].map((t) => ({ ...t, type: "Shared" }));
