+++ src/js/resource/texture/configs/lightning-barrier.ts
@@ -7,4 +7,4 @@ export const LightningBarrierTextureConfigs: TextureConfig[] = [
     id: TEXTURE_IDS.BARRIER_LIGHTNING,
     path: "barrier/lightning/lightning.webp",
   },
-];
+].map((t) => ({ ...t, type: "Shared" }));
