+++ src/js/resource/texture/configs/lightning-dozer.ts
@@ -71,4 +71,4 @@ export const LightningDozerTextureConfigs: TextureConfig[] = [
     id: TEXTURE_IDS.LIGHTNING_DOZER_BOW,
     path: "armdozer/lightning-dozer/bow.webp",
   },
-];
+].map((t) => ({ ...t, type: "Shared" }));
