+++ src/js/resource/texture/configs/predicated-damage.ts
@@ -7,4 +7,4 @@ export const PredicatedDamageTextureConfigs: TextureConfig[] = [
     id: TEXTURE_IDS.PREDICATED_DAMAGE_NUMBER,
     path: "gauge/predicated-damage-number.webp",
   },
-];
+].map((t) => ({ ...t, type: "Shared" }));
