+++ src/js/resource/texture/configs/shock-wave.ts
@@ -11,4 +11,4 @@ export const ShockWaveTextureConfigs: TextureConfig[] = [
     id: TEXTURE_IDS.HITMARK_SHOCK_WAVE_RING,
     path: "hitmark/shock-wave/ring.webp",
   },
-];
+].map((t) => ({ ...t, type: "Shared" }));
