--- src/js/resource/texture/configs/sinya.ts
@@ -1,10 +1,0 @@
-import { TEXTURE_IDS } from "../ids";
-import { TextureConfig } from "../resource";
-
-/** シンヤのテクスチャ設定をあつめたもの */
-export const ShinyaTextureConfigs: TextureConfig[] = [
-  {
-    id: TEXTURE_IDS.SHINYA_CUTIN,
-    path: "pilot/shinya/skill-cutin.webp",
-  },
-];
