+++ src/js/resource/texture/configs/tsubasa.ts
@@ -7,4 +7,4 @@ export const TsubasaTextureConfigs: TextureConfig[] = [
     id: TEXTURE_IDS.TSUBASA_CUTIN,
     path: "pilot/tsubasa/skill-cutin.webp",
   },
-];
+].map((t) => ({ ...t, type: "Shared" }));
