+++ src/js/resource/texture/resource.ts
@@ -1,16 +1,18 @@
 import * as THREE from "three";
 
+import { ResourceType } from "../resource-type";
+
 /** テクスチャID */
 export type TextureId = string;
 
 /** テクスチャ設定 */
-export type TextureConfig = {
+export type TextureConfig = ResourceType & {
   id: TextureId;
   path: string;
 };
 
 /** テクスチャリソース */
-export type TextureResource = {
+export type TextureResource = ResourceType & {
   id: TextureId;
   texture: THREE.Texture;
 };
